fix: Stopped tracking env.local, and env.staging

+++ controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 
 //@desc Get goals
 //@route GET /api/goals
 //@access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user.id });
   res.status(200).json({ goals });
 });
 
@@ -20,6 +21,7 @@ const setGoals = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
@@ -33,6 +35,18 @@ const putGoals = asyncHandler(async (req, res) => {
   if (!goal) {
     res.status(400);
     throw new Error("Goal not found");
+  }
+
+  const user = await User.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    throw new Error("User not authorized");
   }
 
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -51,7 +65,17 @@ const deleteGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user.id);
 
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  //Make sure the logged in user matches the goal user
+  if (goal.user.toString() !== user.id) {
+    throw new Error("User not authorized");
+  }
   const deleteGoal = await Goal.deleteOne({ id: req.params.id });
 
   res.status(200).json(deleteGoal);
